requirements.txt
Flask
mysql-connector-python
requests
pandas
python-binance
schedule
gunicorn